refactor: split upload and download into modules to improve readability
